requirement.txt
asyncio==3.4.3
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
gevent==1.4.0
greenlet==0.4.15
idna==2.8
pyasn1==0.4.7
pycparser==2.19
pycryptodome==3.9.0
requests==2.22.0
rsa==4.0
selenium==3.141.0
urllib3==1.25.3
websocket==0.2.1
websockets==8.0.2
ws4py==0.5.1
